requirements.txt
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.6
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
